fix error

+++ oss-sdk/src/commonMain/kotlin/com/aliyun/kotlin/sdk/service/oss2/models/ModelDownloader.kt
@@ -194,6 +194,8 @@ internal class DownloadCheckpoint(
     }
 
     fun remove() {
-        SystemFileSystem.delete(cpFilePath)
+        if (SystemFileSystem.exists(cpFilePath)) {
+            SystemFileSystem.delete(cpFilePath)
+        }
     }
 }
