+++ oss-sdk/src/commonMain/kotlin/com/aliyun/kotlin/sdk/service/oss2/models/ModelUploader.kt
@@ -198,6 +198,8 @@ internal class UploadCheckpoint(
     }
 
     fun remove() {
-        SystemFileSystem.delete(cpFilePath)
+        if (SystemFileSystem.exists(cpFilePath)) {
+            SystemFileSystem.delete(cpFilePath)
+        }
     }
 }
